refactoring types and constants
