+++ src/router/route-names.ts
@@ -4,11 +4,10 @@
 export enum RouteName {
   DASHBOARD = 'Dashboard',
   DATA = 'Data',
-  INSPECT = 'Inspect',
-  CREATE_NEW = 'CreateNew',
-  CREATE_FOR_ID = 'CreateForId',
-  EDIT = 'Edit',
-  CHARTS = 'Charts',
+  ACTION_INSPECT = 'Inspect',
+  ACTION_CREATE = 'Create',
+  ACTION_EDIT = 'Edit',
+  ACTION_CHARTS = 'Charts',
   SETTINGS = 'Settings',
   FAQ = 'FAQ',
   ABOUT = 'About',
