+++ src/views/ActionChartsView.vue
