+++ src/views/ActionInspectView.vue
