--- src/views/CreateForIdView.vue
@@ -1,12 +1,0 @@
-<script setup lang="ts">
-import ResponsivePage from '@/components/ResponsivePage.vue'
-import { Icon } from '@/constants/icons'
-</script>
-
-<template>
-  <ResponsivePage :banner-icon="Icon.DONATE" banner-title="Donate">
-    <QCard>
-      <QCardSection> Create For Id - PLACEHOLDER </QCardSection>
-    </QCard>
-  </ResponsivePage>
-</template>
